Make bottom navigation
The bottom navigation menu is linked to the current location path. This is done to ensure that the behavior is consistent when using the back button. I commented out the background color property in theme.js for now, because otherwise the bottom navigation color is almost identical to the background making it hard to discern. Further, I added material-ui icons as a dependency to use the (placeholder) icons in the bottom navigation.