Add action button with mode for autoscrolling in transcript
Floating action button that allows mode switches, either autoscroll on or off. It uses wheel and touch events to leave autoscroll mode because scroll events don't work as we use that to autoscroll. There are other scrolling options such as scroll bar or arrow keys that are currently not captured.