Fix another bug in search and a bit of code clean up
There was another bug pertaining to searchIndex. After searching, when clicking away while the search value remains, upon return to search field and pressing enter will lead to the old focus highlight to stay because the searchIndex is reset to -1. Fixed this by introducing a state for searchIndex in combination with useEffect.